The character position was not correct with the Cocoa backend. (see https://sft.its.cern.ch/jira/browse/ROOT-6561)

+++ graf2d/graf/src/TText.cxx
@@ -527,7 +527,7 @@ void TText::GetTextAscentDescent(UInt_t &a, UInt_t &d, const char *text) const
       const Font_t oldFont = gVirtualX->GetTextFont();
       if (gVirtualX->InheritsFrom("TGCocoa"))
          gVirtualX->SetTextFont(fTextFont);
-      gVirtualX->SetTextSize((int)tsize);
+      gVirtualX->SetTextSize(tsize);
       a = gVirtualX->GetFontAscent(text);
       if (!a) {
          UInt_t w;
@@ -559,7 +559,7 @@ void TText::GetTextAscentDescent(UInt_t &a, UInt_t &d, const wchar_t *text) cons
       a = TTF::GetBox().yMax;
       d = TMath::Abs(TTF::GetBox().yMin);
    } else {
-      gVirtualX->SetTextSize((int)tsize);
+      gVirtualX->SetTextSize(tsize);
       a = gVirtualX->GetFontAscent();
       if (!a) {
          UInt_t w;
@@ -591,7 +591,7 @@ void TText::GetTextExtent(UInt_t &w, UInt_t &h, const char *text) const
       const Font_t oldFont = gVirtualX->GetTextFont();
       if (gVirtualX->InheritsFrom("TGCocoa"))
          gVirtualX->SetTextFont(fTextFont);
-      gVirtualX->SetTextSize((int)tsize);
+      gVirtualX->SetTextSize(tsize);
       gVirtualX->GetTextExtent(w, h, (char*)text);
       if (gVirtualX->InheritsFrom("TGCocoa"))
          gVirtualX->SetTextFont(oldFont);
@@ -626,7 +626,7 @@ void TText::GetTextAdvance(UInt_t &a, const char *text, const Bool_t kern) const
       if (gVirtualX->InheritsFrom("TGCocoa"))
          gVirtualX->SetTextFont(fTextFont);
 
-      gVirtualX->SetTextSize((int)tsize);
+      gVirtualX->SetTextSize(tsize);
       gVirtualX->GetTextExtent(a, h, (char*)text);
 
       if (gVirtualX->InheritsFrom("TGCocoa"))
@@ -653,7 +653,7 @@ void TText::GetTextExtent(UInt_t &w, UInt_t &h, const wchar_t *text) const
       TTF::SetTextSize(tsize);
       TTF::GetTextExtent(w, h, (wchar_t*)text);
    } else {
-      gVirtualX->SetTextSize((int)tsize);
+      gVirtualX->SetTextSize(tsize);
       gVirtualX->GetTextExtent(w, h, (wchar_t*)text);
    }
 }
